IGNITE-23128 / move classes to the table module
